🚚 move CommandResultError into errors dir
