+++ src/other/parsing/parseFlags.ts
@@ -3,7 +3,7 @@ import { parseValue } from "./parseValue";
 import { Message } from "discord.js";
 import { ParsableDefinition } from "../../models/definition/ParsableDefinition";
 import { CommandResultUtils } from "../../models/CommandResult";
-import { CommandResultError } from "../../models/CommandResultError";
+import { CommandResultError } from "../../models/errors/CommandResultError";
 
 /** @internal */
 export function parseFlags(
